ML18 code review comment resolution

+++ simulation-service/src/main/java/com/sabara/model/resource/BattleResults.java
@@ -10,7 +10,7 @@ public class BattleResults {
     private BattleType battleType;
     private String winner;
     private List<String> survivors;
-    private int battleDuration;
+    private long battleDuration;
     private BattleMap map;
 
 }
